Rudi->Navi Changed main to be a function (mainLoop) in a class called Navi Updated functions to use the variables from the class where possible
