+++ Bebop2cpp-master/SNAP/navi.cpp
@@ -19,21 +19,127 @@
 #include <fstream>
 #include <iostream>
 #include <sstream>
-#include "SNAP_rudi.h"
+#include "navi.h"
 
 using namespace std;
 
-const double DistAndDirec::INTERSECTION_THRESHOLD = 4.0;
-const double DistAndDirec::WALL_THRESHOLD = 1.0;
+const double Navi::INTERSECTION_THRESHOLD = 4.0;
+const double Navi::WALL_THRESHOLD = 1.0;
 //arbitrary value... probably something better could be used
-double DistAndDirec::INTERSECTION_INDICATOR[2] = { 100000000000, -100000000000 };
+double Navi::INTERSECTION_INDICATOR[2] = { 100000000000, -100000000000 };
 
-//drone position (should only be one copy)
-vector <double*> DistAndDirec::dpos; 
-//which way the drone was moving (again, only ever one copy)
-vector <char> DistAndDirec::ddirec;
+Navi::Navi(){
+	getting_info_loop = true;
+	//node info array
+	nia = NULL;
+	current_direction = 'n';
+	initial_direction = 'n';
+	
+	currently_reversing = false;
+	did_reverse = false;
+}
 
-void handleIntersection( DistAndDirec& intersection, char& current_direction, double* nia_position, double& mbd, double& mfd, double& mld, double& mrd ){
+void Navi::mainLoop( string file_name ){
+	while(getting_info_loop)
+	{	
+		ReadyToReadScan();
+		
+		//nia[0] = node coordinates
+		//nia[1] = max backward distance vector
+		//nia[2] = max forward distance vector
+		//nia[3] = max left distance vector
+		//nia[4] = max right distance vector
+		nia = ReadScanFile(file_name);
+		
+		//helps readability, hopefully..
+		
+		nia_position = nia[0];
+		mbd = nia[1][0];
+		mfd = nia[2][0];
+		mld = nia[3][1];
+		mrd = nia[4][1];
+		
+		if( initial_direction == 'n' )
+		{
+			drone_dd_vals = FindMaxDistance(mbd, mfd, mld, mrd);
+			initial_direction = drone_dd_vals.direction;
+			current_direction = initial_direction;
+			cout << "Go: " << initial_direction << endl;	
+			continue;
+		}
+		
+		intersection = CheckForIntersections(mbd, mfd, mld, mrd, current_direction);		
+		
+		// n indicates no value was set, thus it isn't an intersection
+		if( intersection.direction != 'n')
+		{	
+			handleIntersection();
+		}
+		//need to insert the position after checking the intersection, since
+		// the intersection check looks for it inside. Not a huge deal,
+		// but will be easier since we don't need it til now anyways...
+		// That was a lot of comments for no real good reason.
+		dpos.push_back(nia_position);
+		
+		current_distance = GetDistance(current_direction);
+
+		if( current_distance <= Navi::WALL_THRESHOLD )
+		{
+			cout << "WALLLLLLL!" << endl;
+			current_direction = ReverseDirection(current_direction);
+			currently_reversing = true;
+			dpos.pop_back();
+			//want to remove last position, since we need to reverse back
+			//dpos.pop_back();
+		}
+		
+		if( currently_reversing == true )
+		{
+			cout << "Begin reversing" << endl;
+			vector <double*> :: reverse_iterator dprit;
+			for (dprit = dpos.rbegin(); *dprit != Navi::INTERSECTION_INDICATOR; ++dprit)
+			{
+				ReadyToReadScan();
+				cout << "Go: " << current_direction << endl;
+				//dpos.pop_back();
+			}
+			
+			cout << "End reversing" << endl;	
+			currently_reversing = false;
+			did_reverse = true;
+		}
+	
+		/*
+		vector <double*> :: iterator dpit;	
+		for(dpit = dpos.begin(); dpit != dpos.end(); dpit++)
+		{	
+			cout << *dpit[0] << endl;
+		}	
+
+
+		vector <char> :: iterator ddit;
+		for( ddit = ddirec.begin(); ddit != ddirec.end(); ddit++)
+		{
+			cout << *ddit << endl;
+		}
+		*/
+	
+		//just don't want to print out what direction to go twice	
+		if(did_reverse != true)
+		{
+			cout << "Go: " << current_direction << endl;	
+		}
+		else //we did just reverse
+		{
+			did_reverse = false;
+		}
+	
+	}
+
+
+}
+
+void Navi::handleIntersection(){
 	cout << "INTERSECTION!!!!!" << endl;
 			
 	bool prev_intersection = false;
@@ -46,10 +152,10 @@ void handleIntersection( DistAndDirec& intersection, char& current_direction, do
 	if ( !IsValueInDpos(nia_position) )
 	{
 		cout << "????" << endl;
-		DistAndDirec::ddirec.push_back('x');
+		ddirec.push_back('x');
 		//If we reach an intersection, don't want to go down the way we came
 		// when we come back to it
-		DistAndDirec::ddirec.push_back(ReverseDirection(prev_direction));
+		ddirec.push_back(ReverseDirection(prev_direction));
 	}
 	else //we are at a previous intersection
 	{
@@ -73,7 +179,7 @@ void handleIntersection( DistAndDirec& intersection, char& current_direction, do
 		// went to at this intersection and make their max 0 (temporarily)
 		// when we do this, we know which new direction to go
 		vector <char> :: reverse_iterator ddrit;
-		for (ddrit = DistAndDirec::ddirec.rbegin(); ddrit != DistAndDirec::ddirec.rend() && (*ddrit) != 'x'; ddrit++)
+		for (ddrit = ddirec.rbegin(); ddrit != ddirec.rend() && (*ddrit) != 'x'; ddrit++)
 		{
 			cout << (*ddrit) << endl;
 			ChangeDistance(*ddrit, 0, idist);
@@ -106,17 +212,17 @@ void handleIntersection( DistAndDirec& intersection, char& current_direction, do
 
 	}
 			
-	DistAndDirec::dpos.push_back(DistAndDirec::INTERSECTION_INDICATOR);
+	dpos.push_back(Navi::INTERSECTION_INDICATOR);
 			
 	//don't want to go back down this way when we travel back to the
 	// intersection
-	DistAndDirec::ddirec.push_back(current_direction);
+	ddirec.push_back(current_direction);
 }
 
 /*
 	This just makes sure the user is ready
 */
-void ReadyToReadScan()
+void Navi::ReadyToReadScan()
 {
 	char scan_ready;
 	cout << "Ready for next scan (y for yes; e for exit)" << endl;
@@ -140,7 +246,7 @@ void ReadyToReadScan()
 	}
 }
 
-double **  ReadScanFile(string fn)
+double ** Navi::ReadScanFile(string fn)
 {
 	ifstream fin;
 	double x_info;
@@ -228,7 +334,7 @@ double **  ReadScanFile(string fn)
 	check_x_or_y means which index it should be checking in the array:
 		0 for x, 1 for y
 */
-double * CheckMaxCoord(double x, double y, int check_x_or_y, double current_min[])
+double * Navi::CheckMaxCoord(double x, double y, int check_x_or_y, double current_min[])
 {
 	double * new_min_array = new double[2];
 	
@@ -268,15 +374,13 @@ double * CheckMaxCoord(double x, double y, int check_x_or_y, double current_min[
 	return new_min_array;
 }
 
-DistAndDirec FindMaxDistance(double mbd, double mfd, double mld, double mrd)
+DistAndDirec Navi::FindMaxDistance(double bd, double fd, double ld, double rd)
 {	
-	DistAndDirec drone_dd_vals;
-	
 	/*
 		I am putting mfd first because this would be the direction to move
 		if there is a tie between the distances.
 	*/
-	double distance_array[4] = {mfd, mbd, mld, mrd};
+	double distance_array[4] = {fd, bd, ld, rd};
 
 	double cur_max = 0;
 	int best_index = -1;
@@ -321,43 +425,39 @@ DistAndDirec FindMaxDistance(double mbd, double mfd, double mld, double mrd)
 /*
 	cd = current direction
 */
-DistAndDirec CheckForIntersections(double mbd, double mfd, double mld, double mrd, char cd)
+DistAndDirec Navi::CheckForIntersections(double bd, double fd, double ld, double rd, char cd)
 {
-	DistAndDirec drone_dd_vals;	
-	DistAndDirec intersection_vals;	
-	
-
-	intersection_vals.distance = 0;
-	intersection_vals.direction = 'n';
+	intersection.distance = 0;
+	intersection.direction = 'n';
 	
 	//We check for f and b here, because we don't want to go in the opposite direction
 	// and loop between the two largest values.
 	if( cd == 'f' || cd == 'b' )
 	{
 		//0, 0 is because we only want to look for mld and mrd
-		drone_dd_vals = FindMaxDistance(0, 0, mld, mrd);
-		if (drone_dd_vals.distance >= DistAndDirec::INTERSECTION_THRESHOLD)
+		drone_dd_vals = FindMaxDistance(0, 0, ld, rd);
+		if (drone_dd_vals.distance >= Navi::INTERSECTION_THRESHOLD)
 		{
-			intersection_vals.distance = drone_dd_vals.distance;
-			intersection_vals.direction = drone_dd_vals.direction;
+			intersection.distance = drone_dd_vals.distance;
+			intersection.direction = drone_dd_vals.direction;
 		}
 	}
 
 	if( cd == 'l' || cd == 'r' )
 	{
 		//0, 0 is because we only want to look for mld and mrd
-		drone_dd_vals = FindMaxDistance(mbd, mfd, 0, 0);
-		if (drone_dd_vals.distance >= DistAndDirec::INTERSECTION_THRESHOLD)
+		drone_dd_vals = FindMaxDistance(bd, fd, 0, 0);
+		if (drone_dd_vals.distance >= Navi::INTERSECTION_THRESHOLD)
 		{
-			intersection_vals.distance = drone_dd_vals.distance;
-			intersection_vals.direction = drone_dd_vals.direction;
+			intersection.distance = drone_dd_vals.distance;
+			intersection.direction = drone_dd_vals.direction;
 		}
 	}
 
-	return intersection_vals;
+	return intersection;
 }
 
-double GetDistance(double mbd, double mfd, double mld, double mrd, char cd)
+double Navi::GetDistance(char cd)
 {
 	double current_distance;
 
@@ -387,7 +487,7 @@ double GetDistance(double mbd, double mfd, double mld, double mrd, char cd)
 /*
 	cd = current_direction
 */
-char ReverseDirection(char cd)
+char Navi::ReverseDirection(char cd)
 {
 	//new direction
 	char nd;
@@ -420,7 +520,7 @@ char ReverseDirection(char cd)
 	cv = changed value
 */
 //void ChangeDistance(char cd, double cv, double *mbd, double *mfd, double *mld, double *mrd)
-void ChangeDistance(char cd, double cv, double * dist)
+void Navi::ChangeDistance(char cd, double cv, double * dist)
 {
 	//new value
 	double nv;
@@ -447,10 +547,10 @@ void ChangeDistance(char cd, double cv, double * dist)
 	//cout << cv << (*mfd) << (*mbd) << (*mld) << (*mrd) << endl;	
 }
 
-bool IsValueInDpos(double *value)
+bool Navi::IsValueInDpos(double *value)
 {
 	vector <double*> :: iterator dpit;
-	for( dpit = DistAndDirec::dpos.begin(); dpit != DistAndDirec::dpos.end(); dpit++)
+	for( dpit = dpos.begin(); dpit != dpos.end(); dpit++)
 	{
 		cout << (*dpit)[0] << " " << (*dpit)[1] << " " << value[0] << " " << value[1] << endl;
 		if( (*dpit)[0] == value[0] && (*dpit)[1] == value[1] )
